% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bacteria}
\alias{Bacteria}
\title{Bacteria: Two-way Balanced Factorial Design}
\format{
A data frame with 36 rows and 3 columns:
\describe{
\item{Building}{Building name}
\item{Location}{Location in each building}
\item{Count}{Colony-forming units: a measure of bacteria level}
}
}
\usage{
Bacteria
}
\description{
Two students, Isaac and Courtney, sampled surfaces around their campus to analyze the preva- lence of bacteria. They compared different types of surfaces in both residential and academic buildings. Data were collected by wiping surfaces with a wet Q-tip and swabbing the result on a standard
nutrient agar plate. Locations were all tested on April 24, 2009 over the course of a two-hour period. The plates were incubated at 37°C for 48 hours before colony-forming units (CFUs) were counted as a measure of bacteria levels. Figure 4.5 shows several of the plates from this study. When CFUs exceeded 400 per plate, one fourth of the plate was counted and the total was calculated from that sample.
}
\keyword{datasets}
