% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C5Popcorn}
\alias{C5Popcorn}
\title{C5Popcorn: Block Design}
\format{
An object of class \code{data.frame} with 12 rows and 5 columns.
}
\usage{
C5Popcorn
}
\description{
Two students in a design of experiments course wanted to test if the price and the storage location
of popcorn influenced the percentage of kernels that popped. The students purchased three boxes
of both an expensive and a generic popcorn brand (labeled Exp and Gen). Each box contained six
microwavable bags of popcorn. Two bags were randomly selected from each box and stored for one week,
one in the refrigerator (Frig) and the other at room temperature (Room).
The bags were popped in random order according to the instructions on the box, and
the percentage of popped kernels was calculated for each bag.
}
\keyword{datasets}
