% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Colors2}
\alias{Colors2}
\title{Colors2: Split-Plot Design}
\format{
An object of class \code{data.frame} with 248 rows and 6 columns.
}
\usage{
Colors2
}
\description{
In addition to the 16 females tested in the 'Colors' dataset, 16 males were also tested in this dataset.
}
\keyword{datasets}
