% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cookies}
\alias{Cookies}
\title{Cookies: Split-Plot Design}
\format{
An object of class \code{data.frame} with 135 rows and 4 columns.
}
\usage{
Cookies
}
\description{
Two students wanted to determine if people could taste the difference in chocolate chip cookies
with varying amounts of sugar and varying amount of freshness. Nine batches were made, following the recipe
on the chocolate chip bag as closely as possible except for the amount of sugar.
Each batch was randomly assigned to one of three treatments: half the suggested amount of sugar,
the suggested amount of sugar, and double the suggested amount of sugar.
}
\details{
On the day the nine batches of cookies
were baked, the researchers handed out five cookies from each batch (a total of 45 cookies)
to people in their dorm and asked them to rate the cookies. from 1 through 10, with 1 being inedible and 10
being the best cookie they every had. The research- ers stored the rest of the cookies for a day.
On the second day, the researchers handed out five more cookies from each of the original nine batches to
students in their dorm and asked them to rate them from 1 through 10. The researchers did the same thing
on the third day.

Split-plot designs are often used when time is a second factor.
The whole-plot factor (Sugar) is randomly assigned to whole-plot units (Batch), and
then these same units (Batches) are measured at several time points (Day).
This is called a split plot in time, as the split plots are the time points within the units.

Note that the factor Day is confounded with any other effect that occurs over time.
For example, suppose this study was conducted on a Saturday, Sunday, and Monday.
Students may have been more stressed on Monday and unknowingly tended to give lower scores on Monday.
Or more parents may have been around on the weekend and may have been more positive than students when
rating the cookies.
}
\keyword{datasets}
