% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cookies2}
\alias{Cookies2}
\title{Cookies2: Block Design}
\format{
An object of class \code{data.frame} with 144 rows and 5 columns.
}
\usage{
Cookies2
}
\description{
Two students wanted to test whether ingredients (butter, Fleischmann’s corn oil margarine,
or unflavored Crisco), cooking time (short or long) or cookie type (chocolate chip or gingersnap)
influenced taste ratings. Both main effects and interactions were of interest.
Twelve volunteers were found who were each willing to taste
12 cookies in random order (one taste for each of the two cookie types,
the three ingredient types, and the two cooking times).
Each volunteer ranked all 12 cookies on a scale from 1 to 10 (10 being the best).
}
\keyword{datasets}
