% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Corn}
\alias{Corn}
\title{Corn: Split-Plot Design}
\format{
An object of class \code{data.frame} with 40 rows and 5 columns.
}
\usage{
Corn
}
\description{
Many seed companies and state-sponsored agricultural research centers have large testing areas
that have been shown to be similar to other plots of land throughout their state. In a large testing
area, eight 20-acre plots of land were randomly assigned to a nitrogen application rate (0, 70, 140,
or 210 pounds of nitrogen per acre). Each of these 20-acre plots was also subdivided into five
4-acre subplots. Within each 20-acre plot, the subplots were assigned to be planted with one species
of hybrid corn (A, B, C, D, or E). At the end of the season, the fields were harvested and the yield
in bushels per acre was recorded.
}
\keyword{datasets}
