% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cups}
\alias{Cups}
\title{Cups: Fractional Factorial Design}
\format{
An object of class \code{data.frame} with 32 rows and 7 columns.
}
\usage{
Cups
}
\description{
A manager of a manufacturing company, shift managers, cup machine operators, and a statistician
decided to identify which factors were most influential in keeping their cups from leaking.
Over 30 possible factors were identified, but after some thoughtful discussions the group settled
on six variables that should be tested for their effects on leaking cups.
}
\keyword{datasets}
