% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_aov}
\alias{dox_aov}
\title{An ANOVA summary table with total DF & SS}
\usage{
dox_aov(aov)
}
\arguments{
\item{aov}{the anova model}
}
\value{
a html table
}
\description{
This function gives an ANOVA summary table with total degrees of freedom and sum of squares
}
