% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_boxplot}
\alias{dox_boxplot}
\title{A box plot to check the equal variances assumption in ANOVA}
\usage{
dox_boxplot(formula, dataset, color = NULL, facet = NULL)
}
\arguments{
\item{formula}{y~x1}

\item{dataset}{the dataset that contains the experiment information}

\item{color}{colored by this factor (optional)}

\item{facet}{faceted by this factor (optional)}
}
\value{
a box plot
}
\description{
This function gives a box plot to check the within groups equal variances assumptions.
The first factor x1 is the x-axis, and if the second factor x2 exists,
the box plot will be colored by x2.
}
