% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_contrast}
\alias{dox_contrast}
\title{Checking for contrasts and comparing between different treatment levels
This function checks for contrasts between a pair or more set of different treatment level configurations
mosaic library shoudl be installed and loaded
dox_contrast(Time~Bulb, dataset=Lightbulb)}
\usage{
dox_contrast(formula, dataset, alpha = 0.05, method = "All")
}
\description{
Checking for contrasts and comparing between different treatment levels
This function checks for contrasts between a pair or more set of different treatment level configurations
mosaic library shoudl be installed and loaded
dox_contrast(Time~Bulb, dataset=Lightbulb)
}
