% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_flextable}
\alias{dox_flextable}
\title{Flextable generator}
\usage{
dox_flextable(formula, dataset)
}
\arguments{
\item{formula}{y~x}

\item{dataset}{the dataset that contains the experiment information}
}
\value{
a flextable
}
\description{
This function gives a flextable to visualize the result
}
\examples{
flextable(LogStrength ~ Brand, Towels2)

}
