% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{dox_main}
\alias{dox_main}
\title{Main Effect Plots}
\usage{
dox_main(
  formula,
  dataset,
  label = "Mean",
  text_size = text_size,
  ylim,
  title1 = NULL,
  title2 = NULL,
  title3 = NULL,
  title4 = NULL
)
}
\arguments{
\item{formula}{y~x1+x2(optional)+x3(optional)}

\item{dataset}{the dataset that contains the experiment information}

\item{label}{Label "Mean" or "Effect"}

\item{text_size}{text size of the x-axis}

\item{ylim}{a vector of the range of the y-axis}
}
\value{
main effect plots
}
\description{
This function gives the main effect plots for one or more treatments in an experiment.
If more than 1 treatment is given and ylim is not specified, plots will share the same y-axis range.
}
\examples{
dox_main(LogStrength ~ Brand + Water, Towels2, label="Mean")
# If you want the label to be effect and have a larger size for the x-axis
dox_main(LogStrength ~ Brand + Water, Towels2, label="Effect", text_size = 14,title1="default")
}
