% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_pairs}
\alias{dox_pairs}
\title{Confidence intervals of pairwise comparisons}
\usage{
dox_pairs(formula, dataset, alpha = 0.05, method = "ALL")
}
\arguments{
\item{formula}{target~treatment}

\item{dataset}{dataset of experimental results}

\item{alpha}{alpha level (default 0.05)}

\item{method}{LSD, BSD, or HSD (default is ALL)}
}
\value{
confidence interval plots
}
\description{
This function plots the confidence intervals of pairwise comparisons using Fisher least significant difference (LSD),
Bonferroni significant difference (BSD), and Tukey honest significant difference (HSD) methods.
}
\examples{
dox_pairs(LogStrength~Water, Towels2)
# If you want to adjust the alpha level
dox_pairs(LogStrength~Water, Towels2, alpha = 0.01)
# If you are only interested in LSD
dox_pairs(LogStrength~Water, Towels2, method = "LSD")
}
