% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_resid}
\alias{dox_resid}
\title{Residual vs fit/order plots}
\usage{
dox_resid(anova_model, dataset, plot = "all", bins = 30)
}
\arguments{
\item{anova_model}{the anova model}

\item{dataset}{the dataset that contains the experiment information}

\item{plot}{which of the four plots to show. Default is to show all four}

\item{bins}{the number of bins in histogram}
}
\value{
A qqplot and a histogram for residuals; a residual versus fit plot and a residual versus order plot
}
\description{
This function gives four plots to check the "independent and identically distributed observations"
and "normality" assumption in ANOVA. 1.a qqplot for residuals; 2.and a histogram for residuals;
3.residual versus fit plot; 4.residual versus order plot
}
