% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_scatterplot}
\alias{dox_scatterplot}
\title{A scatterplot to check the equal variances assumption in ANOVA}
\usage{
dox_scatterplot(formula, dataset, color = NULL, facet = NULL, jitter = FALSE)
}
\arguments{
\item{formula}{y~x}

\item{dataset}{the dataset that contains the experiment information}

\item{color}{colored by this factor (optional)}

\item{facet}{faceted by this factor (optional)}

\item{jitter}{the width of jitter or FALSE (not to use jitter)}
}
\value{
a scatterplot
}
\description{
This function gives a scatterplot to check the within groups equal variances assumptions.
}
\examples{
dox_scatterplot(LogStrength ~ Brand, Towels2, color = Water)
# if you want to use jitter
dox_scatterplot(LogStrength ~ Brand, Towels2, color = Water, jitter = 0.05)

# if you want to use facet
dox_scatterplot(LogStrength ~ Brand, Towels2, facet = Water, jitter = 0.15)
dox_scatterplot(LogStrength ~ Brand, Towels2, facet = Brand~Water, jitter = 0.15)
}
