% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_split_aov}
\alias{dox_split_aov}
\title{An ANOVA summary table specifically for SPlit Plot designs}
\usage{
dox_split_aov(formula, dataset)
}
\arguments{
\item{formula}{formula used in ANOVA}

\item{dataset}{the dataset that contains the experiment information}
}
\value{
a html table and residuals,fitted values
}
\description{
This function gives an ANOVA summary table with total degrees of freedom and sum of squares
}
\examples{
dox_split_aov(LogStrength ~ Brand + Water, Towels2)
}
