% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_sumstat}
\alias{dox_sumstat}
\title{Summary Statistics}
\usage{
dox_sumstat(formula, dataset)
}
\arguments{
\item{formula}{y~x1+x2}

\item{dataset}{the dataset that contains the experiment information}
}
\value{
A table that gives summary statistics about y partitioned by treatments
}
\description{
This function applies the favstats function in mosaic package and gives summary statistics.
}
