% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_table}
\alias{dox_table}
\title{An interactive table to show variances of different groups}
\usage{
dox_table(dataset, response, x1, x2 = NULL, x3 = NULL)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor (optional)}

\item{x3}{the third factor (optional)}
}
\value{
an interactive variance table
}
\description{
This function gives an interactive table to show variances of different groups.
It can help check the within groups equal variances assumptions. You can click column names to sort.
}
