% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Flower}
\alias{Flower}
\title{Flower: Block Design}
\format{
An object of class \code{data.frame} with 18 rows and 4 columns.
}
\usage{
Flower
}
\description{
Students in an introductory statistics class tested the impact of different types of
water solutions on the lon- gevity of cut flowers. They purchased 18 white carnations from three different
randomly selected  stores and randomly assigned each flower to one of three treatments
(plain water, one aspirin crushed and added
to the water, and a floral compound provided by the flower shop) and then measured how many days
it took until the flower wilted.
}
\keyword{datasets}
