% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Football}
\alias{Football}
\title{Football: Split-Plot Design
Ben, Hugh, and Alex wanted to determine if the size of the “football” made a
difference in scoring accuracy in the game of paper football. (Details of the simple
tabletop game can be found at http:// www.paperfootball.us.) In addition, these students
were interested in knowing if the effect of football size was dependent on a player’s experience level.}
\format{
An object of class \code{data.frame} with 36 rows and 5 columns.
}
\usage{
Football
}
\description{
The researchers set up a goal that was 8 inches above a table and 10 inches wide.
The subjects kicked/flicked the football from 42 inches away from the goal.
}
\details{
There were 18 volunteers, who self-identified as experienced or inexperienced players.
After a few practice kicks, each subject kicked 20 small and 20 large footballs in random order
(flipping a coin before each kick) at the goal. The response for this study was the proportion
of successful goals.
}
\keyword{datasets}
