% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Handwash}
\alias{Handwash}
\title{Handwash: Block Design}
\format{
An object of class \code{data.frame} with 30 rows and 4 columns.
}
\usage{
Handwash
}
\description{
Two students are interested in the effectiveness of three types of cleansers. They also sort CFU counts
into three blocks (representing low, medium, and high CFU counts).
}
\keyword{datasets}
