% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{iid}
\alias{iid}
\title{Residual vs fit/order plots}
\usage{
iid(dataset, anova_model)
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}

\item{anova_model}{the anova model}
}
\value{
A residual versus fit plot and a residual versus order plot
}
\description{
This function gives a residual versus fit plot and a residual versus order plot to check the
"independent and identically distributed observations" assumption in ANOVA
}
