% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Memory}
\alias{Memory}
\title{Memory: Split-Plot Design}
\format{
An object of class \code{data.frame} with 48 rows and 8 columns.
}
\usage{
Memory
}
\description{
Two students in a statistics course, Josh and Ann, were interested in conducting an experiment
to help them better understand memory. For each test, Josh and Ann decided to test what psychologists
call free recall by asking each subject to read through a list of 20 words for 30 seconds,
work through a distracter, and then recall as many words as he or she could.
Josh and Ann asked each subject to take all four tests in random order, with a short break between tests.
}
\details{
In this experiment, each unit is a test that is randomly assigned to one of the four factor-level combinations listed below:
\enumerate{
\item Abstract word list and mathematics distracter
\item Abstract word list and poetry distracter
\item Concrete word list and mathematics distracter
\item Concrete word list and poetry distracter
}

Also, students are randomly selected from four different majors.
}
\keyword{datasets}
