% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MemoryA}
\alias{MemoryA}
\title{MemoryA: Two-way Balanced Factorial Design}
\format{
A data frame with 100 rows and 4 columns:
\describe{
\item{Age}{Young or old people}
\item{Process}{Instruction types described anove}
\item{Words}{Number of words remembered}
\item{sqrtWords}{Square root of number of words remembered}
}
}
\usage{
MemoryA
}
\description{
Michael Eysenck tested 100 subjects (50 people between the ages of 55 and 65, 50 younger people)
to determine if there was a relationship between age and memory.
10 Each subject was shown 27 words and asked to recall as many of those words as possible.
He also tested whether five different techniques impacted memory.
Each subject was given one of five types of instructions:
\describe{
\item{Counting}{count the number of letters in each word}
\item{Rhyming}{think of a word that rhymes with each word}
\item{Adjective}{think of an adjective to describe each word}
\item{Imagery}{create an image of each word}
\item{Intentional}{remember as many words as possible}
}
The subjects in the first four groups were not aware that they would later be asked to recall each word.
}
\keyword{datasets}
