% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Movies}
\alias{Movies}
\title{Movies: Two-way Balanced Factorial Design}
\format{
A data frame with 210 rows and 4 columns:
\describe{
\item{Movie}{Movie name}
\item{Total.Gross}{Gross earnings}
\item{Genre}{Genre of the movie}
\item{Rating}{Rating of the movie}
}
}
\usage{
Movies
}
\description{
The file Movies contains the ratings and genres of movies that came out in 2008.
This is simply an observational study, as clearly movie producers
do not try to create an equal number of  G, PG, and R movies each year.
Note:
\describe{
\item{G}{Appropriate for people of all ages.}
\item{PG-13}{Parents Strongly Cautioned under the age of 13.}
\item{R}{Children Under 17 Require Accompanying Parent or Adult Guardian.}
}
}
\keyword{datasets}
