% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Music}
\alias{Music}
\title{Music: Block Design}
\format{
An object of class \code{data.frame} with 160 rows and 9 columns.
}
\usage{
Music
}
\description{
Students designed an experiment testing whether music tempo or length of test (1 minute or 3 minutes)
influenced students’ ability to type fast and accurately. Would subjects listening to Stayin’ Alive type
at a different speed than subjects listening to Yesterday by the Beatles?
}
\details{
Forty undergraduate students consented to be in the study. Each subject took four tests from the website
typingtest.com in random order based on two coin flips: 1 min/Yesterday, 1 min/Stayin’ Alive,
3 min/Yesterday, and 3 min/Stayin’ Alive. The questions the researcher wanted to test were the effect of Song,
Length, and Song\emph{Length on words per minute (WPM). The Music dataset also records Accuracy and helps compute
average WPM for each subject (Subject.WPM.Ave), difference of the WPM of each trial and Subject.WPM.Ave,
and Accuracy}WPM (AWPM).
}
\keyword{datasets}
