% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{normal_err}
\alias{normal_err}
\title{A qqplot and a histogram for residuals}
\usage{
normal_err(
  dataset_residual,
  response,
  x1,
  x2 = NULL,
  interaction = TRUE,
  bins = 30
)
}
\arguments{
\item{dataset_residual}{the dataset that contains the experiment information or the residuals}

\item{response}{the value of interest}

\item{x1}{the first factor}

\item{x2}{the second factor (optional)}

\item{interaction}{whether interaction x1*x2 is considered}

\item{bins}{the number of bins}
}
\value{
A qqplot and a histogram for residuals
}
\description{
This function gives a qqplot and a histogram for residuals to check the normality assumption of ANOVA.
You can either use the residuals as the parameters: normal_err(residuals, bins = 30) or you can use a similar pattern
as other functions normal_err(dataset, response, ...).
}
