% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Popcorn}
\alias{Popcorn}
\title{Popcorn Dataset}
\format{
A data frame with 32 rows and 5 columns:
\describe{
\item{Brand}{Brand name}
\item{Microwave}{Microwave location}
\item{Time}{Microwave time}
\item{Unpopped..percent.}{Unpopped percent}
\item{Poprate}{Popped percent}
}
}
\usage{
Popcorn
}
\description{
A dataset containing the pop rate experiments under different conditions
}
\keyword{datasets}
