% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tennis}
\alias{Tennis}
\title{Tennis: Block Design}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\usage{
Tennis
}
\description{
Students on the college tennis team were interested in knowing if string tension
affected the speed and accuracy of a tennis ball.
After warming up, five men’s varsity tennis players volunteered to hit 30 serves using three different
racquet tensions over three days. Day was not treated as a factor of interest in the study,
but used to avoid fatigue in the players. For each serve, players aimed at a target in the back center
of the service box, and accuracy was measured as the distance in inches from the center of the
box to the ball strike location. A radar gun was used to measure the velocity of each player’s serve.
}
\keyword{datasets}
