% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Towels2}
\alias{Towels2}
\title{Towels2: Two-way Balanced Factorial Design}
\format{
A data frame with 156 rows and 4 columns:
\describe{
\item{Brand}{Comfort or Decorator}
\item{Water}{Drops of water}
\item{Strength}{Weights in grams until paper towel breaks}
\item{LogStrength}{Log of the weights in grams until paper towel breaks}
}
}
\usage{
Towels2
}
\description{
Several television advertisements had claimed that a certain brand of paper towel was the strongest,
and these students wanted to determine if there really was a difference.
The students sampled 26 towels from two brands of paper towels, Comfort and Decorator.
Also, they were interested in whether the amount of water impacts the strength of either brand.
Weights (10, 25, 50, 100, or 250 grams) were slowly added to the center of each towel by a third person until it broke.
}
\keyword{datasets}
