% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cholesterol}
\alias{Cholesterol}
\title{Cholesterol: Three-way Balanced Factorial Design}
\format{
A data frame with 72 rows and 5 columns:
\describe{
\item{Patient}{Patient number}
\item{Exercises}{Exercise or not}
\item{Diet}{Diet or not}
\item{Drug}{Types of drugs}
\item{Cholesterol}{Milligrams per deciliter: a measure of cholesterol level}
}
}
\usage{
Cholesterol
}
\description{
Cholesterol is a waxy substance found in blood and cell membranes.
All animals need some cholesterol in their system;
however, too much cholesterol can cause heart attacks and strokes.
A study was conducted to determine how exercise, diet, and three types of drugs impact cholesterol levels.
Seventy-two patients at a nearby hospital who had been diagnosed with high cholesterol
(a level greater than 240 milligrams per deciliter) consented to be in the study.
Each of the 72 patients was randomly assigned to a specific treatment, and after six months the patients’ cholesterol levels were measured again.
}
\keyword{datasets}
