% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_contrast}
\alias{dox_contrast}
\title{Checking for contrasts}
\usage{
dox_contrast(formula, dataset, alpha = 0.05, method = "All")
}
\arguments{
\item{dataset}{the dataset that contains the experiment information}
}
\value{
a html table and residuals,fitted values
}
\description{
Checking for contrasts
}
