% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{dox_flextable}
\alias{dox_flextable}
\title{}
\usage{
dox_flextable(formula, dataset)
}
\arguments{
\item{formula}{formula used in ANOVA}

\item{dataset}{the dataset that contains the experiment information}
}
\value{

}
\description{

}
\examples{
dox_flextable(LogStrength ~ Brand + Water, Towels2)
}

