% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{dox_inter}
\alias{dox_inter}
\title{Interaction Effect Plot}
\usage{
dox_inter(formula, dataset, facet = NULL, label = "Mean", text_size = 12)
}
\arguments{
\item{formula}{y~x1+x2}

\item{dataset}{the dataset that contains the experiment information}

\item{facet}{faceted by this factor (optional)}

\item{label}{Label "Mean" or "Effect"}

\item{text_size}{text size of the x-axis}
}
\value{
interaction effect plot
}
\description{
This function gives the interaction effect plots for two treatments in an experiment.
}
