% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{dox_inter}
\alias{dox_inter}
\title{Interaction Effect Plot}
\usage{
dox_inter(formula, dataset, label = "Mean", text_size = 12)
}
\arguments{
\item{formula}{y~x1+x2}

\item{dataset}{the dataset that contains the experiment information}

\item{label}{Label "Mean" or "Effect"}

\item{text_size}{text size of the x-axis}
}
\value{
interaction effect plot
}
\description{
This function gives the interaction effect plots for two treatments in an experiment.
}
\examples{
dox_inter(LogStrength ~ Brand + Water, Towels2, label="Mean", text_size = 14)
# If you want the label to be effect and have a larger size for the x-axis
dox_inter(LogStrength ~ Brand + Water, Towels2, label="Effect", text_size = 14)
}
