% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_resid}
\alias{dox_resid}
\title{Residual vs fit/order plots}
\usage{
dox_resid(formula, dataset, plot = "all", bins = 30, model_name = "NULL")
}
\arguments{
\item{formula}{formula used in ANOVA}

\item{dataset}{the dataset that contains the experiment information}

\item{plot}{which of the four plots to show. Default is to show all four}

\item{bins}{the number of bins in histogram}

\item{model_name}{model name used in the column names for residuals and fitted values}
}
\value{
A qqplot and a histogram for residuals; a residual versus fit plot and a residual versus order plot
}
\description{
This function gives four plots to check the "independent and identically distributed observations"
and "normality" assumption in ANOVA. 1.a qqplot for residuals; 2.a histogram for residuals;
3.residual versus fit plot; 4.residual versus order plot.
Also, this function invisibly returns the dataset with residuals and fitted values.
}
\examples{
dox_resid(LogStrength~Brand*Water, Towels2)
# If you want to check a specific plot, use plot =
dox_resid(LogStrength~Brand*Water, Towels2, plot = 2, bins = 40)
}
