% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_scatterplot}
\alias{dox_scatterplot}
\title{A scatterplot to check the equal variances assumption in ANOVA}
\usage{
dox_scatterplot(formula, dataset, color = NULL, facet = NULL, jitter = FALSE)
}
\arguments{
\item{formula}{y~x1}

\item{dataset}{the dataset that contains the experiment information}

\item{color}{colored by this factor (optional)}

\item{facet}{faceted by this factor (optional)}

\item{jitter}{whether to use geom_jitter}
}
\value{
a scatterplot
}
\description{
This function gives a scatterplot to check the within groups equal variances assumptions.
The first factor x1 is the x-axis, and if the second factor x2 exists,
the scatterplot will be colored by x2.
}
