% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{dox_table}
\alias{dox_table}
\title{An interactive table to show standard deviations of different groups}
\usage{
dox_table(formula, dataset)
}
\arguments{
\item{formula}{y~x1+x2(optional)+x3(optional)}

\item{dataset}{the dataset that contains the experiment information}
}
\value{
an interactive variance table
}
\description{
This function gives an interactive table to show standard deviations of different groups.
It can help check the within groups equal variances assumptions. You can click column names to sort.
}
